refactor: rename grid components for clarity
