+++ packages/web/src/views/Calendar/components/Grid/AllDayRow/styled.ts
@@ -8,6 +8,8 @@ import {
 } from "@web/views/Calendar/layout.constants";
 import { GRID_TIME_STEP } from "@web/views/Calendar/layout.constants";
 
+import { Columns } from "../Columns/styled";
+
 const gridHeight = `100% - (${GRID_Y_START}px + ${GRID_PADDING_BOTTOM}px)`;
 const gridRowHeight = `(${gridHeight}) / 11`;
 const interval = 60 / GRID_TIME_STEP;
@@ -15,6 +17,9 @@ const allDayRowHeight = `${gridRowHeight} / ${interval}`;
 
 const gridWidth = `100% - ${SCROLLBAR_WIDTH}px`;
 
+export const StyledAllDayColumns = styled(Columns)`
+  height: 100%;
+`;
 export const StyledAllDayRow = styled(Flex)<{ rowsCount: number }>`
   border-bottom: ${({ theme }) => `2px solid ${theme.color.gridLine.primary}`};
   height: ${({ rowsCount }) =>
