--- packages/web/src/views/Calendar/components/Grid/Columns/index.ts
@@ -1,2 +1,0 @@
-import { Columns } from "./Columns";
-export { Columns };
