--- packages/web/src/views/Calendar/components/Grid/MainGrid/GridRows/GridRows.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import { getHourLabels } from "@web/common/utils/web.date.util";
-
-import { StyledGridRows, StyledGridRow } from "./styled";
-
-export const GridRows = () => {
-  return (
-    <StyledGridRows>
-      {getHourLabels().map((dayTime, index) => (
-        <StyledGridRow key={`${dayTime}-${index}:dayTimes`} />
-      ))}
-    </StyledGridRows>
-  );
-};
