--- packages/web/src/views/Calendar/components/Grid/MainGrid/GridRows/index.ts
@@ -1,2 +1,0 @@
-import { GridRows } from "./GridRows";
-export { GridRows };
